in `#save` (without bang), `ActiveRecord::RecordInvalid` error is not passed outward.
Even if no validation error occurs on self, there may be cases where an  `ActiveRecord::RecordInvalid`
exception occurs due to the execution of `#save!` on a different model in callbacks like `after_save`.
In the case of the original ActiveRecord, such occurrences without the bang version of `#save` result
in simply returning false without raising an exception.
Here, we have aligned the behavior with the original ActiveRecord.