Stop implementing `define_attribute_methods` for attributes defined with `delegate_attribute`.
By implementing `define_attribute_methods`, methods such as `_changed?`
are defined for attributes defined with `delegate_attribute` when
subclasses include `ActiveModel::Dirty`, for example.
This is not a problem in itself, but since dirty checking does not
actually work, we decided it would be better not to do this to avoid
confusion.