Secondary database transaction support.
This makes it easier to use in applications with multiple database connections.

One thing to note is that it is difficult to update a model that has multiple connections.
Due to the characteristics of the activerecord API supported by Rails,
a full two-phase commit does not exist.
At the end of the day, nothing more can be done than nested transactions like the one below.