We've simplified the documentation comment yard. (#52)
* Replace ClassMethods with class_methods.

* Map the location of the yard documentation and source code as much as possible.

We had previously consolidated everything into the model class, but we
wanted to define it in a natural way in each location. This was a hack
caused by steep's inability to resolve the issue of `class_methods`, but
by reconsidering our stance toward Steep, we changed it to a more
natural way of writing using `class_methods`.

* setting yard to nicely

* add webrick for yard server command
* always specify --no-private options
* add activesupport-concern plugin